refactor: decompose tui/app.rs into focused state modules (#433)
Extract 4 modules from the 1191-LOC app.rs monolith:
- scroll_controller.rs (73 LOC): standalone ScrollController struct
- app_tools.rs (105 LOC): tool execution state methods
- app_input.rs (119 LOC): input buffer delegation methods
- focus_manager.rs (194 LOC): focus management, LayoutCache, PaneWrapper

Uses Rust's split-impl-block pattern to keep `impl App` blocks in
sibling modules within the same crate. Fields accessed by sibling
modules use pub(super) visibility.

app.rs retains: App struct definition, new(), state queries, message
management, streaming/thinking, autocomplete/modal delegation, and
core state mutations (726 LOC with tests).
